# Optimizations in immutable.Map.{get, contains}