--- a/bincompat-backward.whitelist.conf
+++ b/bincompat-backward.whitelist.conf
@@ -32,6 +32,22 @@ filter {
     {
       matchName="scala.sys.process.ProcessImpl#CompoundProcess.getExitValue"
       problemName=DirectMissingMethodProblem
+    },
+    {
+      matchName="scala.collection.immutable.HashMap.contains0"
+      problemName=DirectMissingMethodProblem
+    },
+    {
+      matchName="scala.collection.immutable.HashMap#HashTrieMap.contains0"
+      problemName=DirectMissingMethodProblem
+    },
+    {
+      matchName="scala.collection.immutable.HashMap#HashMap1.contains0"
+      problemName=DirectMissingMethodProblem
+    },
+    {
+      matchName="scala.collection.immutable.HashMap#HashMapCollision1.contains0"
+      problemName=DirectMissingMethodProblem
     }
   ]
 }