--- a/bincompat-forward.whitelist.conf
+++ b/bincompat-forward.whitelist.conf
@@ -42,6 +42,22 @@ filter {
     {
       matchName="scala.util.hashing.MurmurHash3.wrappedArrayHash"
       problemName=DirectMissingMethodProblem
+    },
+    {
+      matchName="scala.collection.immutable.HashMap.contains0"
+      problemName=DirectMissingMethodProblem
+    },
+    {
+      matchName="scala.collection.immutable.HashMap#HashTrieMap.contains0"
+      problemName=DirectMissingMethodProblem
+    },
+    {
+      matchName="scala.collection.immutable.HashMap#HashMap1.contains0"
+      problemName=DirectMissingMethodProblem
+    },
+    {
+      matchName="scala.collection.immutable.HashMap#HashMapCollision1.contains0"
+      problemName=DirectMissingMethodProblem
     }
   ]
 }