- Avoid allocation of Some in get
  - defer integer left shift until needed
  - avoid redundantly masking an integer

before:

    Benchmark                  (size)  Mode  Cnt      Score      Error  Units
    HashMapBenchmark.contains      10  avgt   20    284.624 ±   18.985  ns/op
    HashMapBenchmark.contains     100  avgt   20   3190.580 ±   33.622  ns/op
    HashMapBenchmark.contains    1000  avgt   20  52967.171 ± 1524.834  ns/op
    HashMapBenchmark.get           10  avgt   20    248.168 ±    2.612  ns/op
    HashMapBenchmark.get          100  avgt   20   2795.469 ±   54.458  ns/op
    HashMapBenchmark.get         1000  avgt   20  52238.773 ± 1268.764  ns/op

after:

    Benchmark                  (size)  Mode  Cnt      Score     Error  Units
    HashMapBenchmark.contains      10  avgt   20    195.107 ±   2.442  ns/op
    HashMapBenchmark.contains     100  avgt   20   2454.151 ±  24.392  ns/op
    HashMapBenchmark.contains    1000  avgt   20  40722.993 ± 520.473  ns/op
    HashMapBenchmark.get           10  avgt   20    245.282 ±   3.547  ns/op
    HashMapBenchmark.get          100  avgt   20   2729.669 ±  32.767  ns/op
    HashMapBenchmark.get         1000  avgt   20  49568.410 ± 794.565  ns/op