@retronym 2017-02-03
@paplorinc I've incorporated your suggestions and re-run the benchmark.