@paplorinc 2017-02-04
Related to: https://github.com/scala/scala/pull/5600