@rorygraves 2017-02-04
@retronym @paplorinc 
Looking at the test failure - its unrelated
```
[info] ! Division of Duration by Long.on overflow edge cases: Exception raised on property evaluation.
[info] > Exception: org.scalacheck.Gen$Choose$IllegalBoundsError: invalid bounds: low=9223372036854775799, high=-9223372036854775797
```
I did some investigation and the failure occurs here:
https://github.com/scala/scala/blob/2.12.x/test/scalacheck/duration.scala#L32
A value of 1 or -1 fails the choose on line 35.
I don't have time to raise the proper bug report right now - but thought I would highlight it.