@retronym 2017-02-04
Noted the unrelated test failure as https://github.com/scala/scala-dev/issues/296