@adriaanm 2017-04-10
Looking over the bytecode diff between 2.12.1 and 2.12.2 (see https://github.com/scala/scala-dev/issues/331), this added a (byte code) public method to a public class. Just noting for the record that this method should not be called from user (Java) code (it's not visible to Scala code outside of the package).

```diff
--- a/library/scala/collection/immutable/HashMap.class
+++ b/library/scala/collection/immutable/HashMap.class
@@ -251,6 +251,34 @@ public class scala/collection/immutable/HashMap extends scala/collection/immutab
<snip>
+  // access flags 0x1
+  // signature (TA;II)Z
+  // declaration: boolean contains0(A, int, int)
+  public contains0(Ljava/lang/Object;II)Z
```